requirements.txt
python-dotenv
langchain
langchain-azure-ai
azure-identity
azure-core
black
langchain-community